Use a different obj directory for the release build
Use obj-...-release for the release binaries so you can build the debug and release
versions without having them clobbering each other.
Disable telemetry (analytics) in the mozconfig.
Disable the crash reporter in the mozconfig, it is not working anyway,
the crash reporter throws an exception when trying to send the report.
Don't build the tests to try to decrease the build time.